+++ BeatTheBox/MFCC.cpp
@@ -13,16 +13,16 @@ template <class T>
 T* MFCC<T>::_H;
 
 template <class T>
-int MFCC<T>::_N;
+size_t MFCC<T>::_N;
 
 template <class T>
-int MFCC<T>::_M; 
+size_t MFCC<T>::_M; 
 
 template <class T>
-int MFCC<T>::_f_min;
+size_t MFCC<T>::_f_min;
 
 template <class T>
-int MFCC<T>::_f_max;
+size_t MFCC<T>::_f_max;
 
 template <class T>
 T MFCC<T>::_deltaPhi; 
@@ -31,34 +31,34 @@ template <class T>
 T MFCC<T>::_fs;
 
 template <class T>
-void MFCC<T>::init(int freqs,int mels, T fs,int minF,int maxF){
+void MFCC<T>::init(size_t freqs,size_t mels, T fs,size_t minF,size_t maxF){
     _N=freqs; _M=mels; _fs=fs; _f_min=minF; _f_max=maxF;
     generateTable();
 }
 
 template <class T>
-void MFCC<T>::init(int spectrumLength) {
+void MFCC<T>::init(size_t spectrumLength) {
     init(spectrumLength, NUM_MELS, 44100, 20, 22050);
 }
 
 template <class T>
-T MFCC<T>::amplitude(int freq, int mel){
+T MFCC<T>::amplitude(size_t freq, size_t mel){
     return _H[freq*_M+mel];
 }
 
 template <class T>
-T* MFCC<T>::getMFCCs(T* spectrum, int spectrumLength){
+T* MFCC<T>::getMFCCs(T* spectrum, size_t spectrumLength){
     if(!_H || spectrumLength!=_N){
         init(spectrumLength);
     }
     
-    for(int i=0;i<spectrumLength;i++){
+    for(size_t i=0;i<spectrumLength;i++){
         assert(spectrum[i]>=0);
     }
     T *mfccs = new T[_M];
-    for(int m=0; m<_M; m++){
+    for(size_t m=0; m<_M; m++){
         T sum = 0;
-        for(int n=0; n<_N; n++){
+        for(size_t n=0; n<_N; n++){
             sum+=_H[n*_M+m] * spectrum[n];
         }
         mfccs[m]=log(1+sum);
@@ -70,35 +70,35 @@ template <class T>
 void MFCC<T>::generateTable(){
     _deltaPhi=(phi(_f_max) - phi(_f_min)) / _M+1;
     _H = new T[_N*_M];
-    for (int i=0; i<_N;i++){
-        for (int j=0; j<_M;j++){
+    for (size_t i=0; i<_N;i++){
+        for (size_t j=0; j<_M;j++){
             _H[i*_M+j]=compH(i+1,j+1);
         }
     }
 }
 
 template <class T>
-T MFCC<T>::phi(int f){
+T MFCC<T>::phi(size_t f){
     return (2595 * log10(f/700+1));
 }
 
 template <class T>
-T MFCC<T>::freq_c(int m){
+T MFCC<T>::freq_c(size_t m){
     return 700*(pow(10,phi_c(m)/2595)-1);
 }
 
 template <class T>
-T MFCC<T>::freq(int k){
+T MFCC<T>::freq(size_t k){
     return k*(_fs/2)/_N;
 }
 
 template <class T>
-T MFCC<T>::phi_c(int m){
+T MFCC<T>::phi_c(size_t m){
     return m * _deltaPhi;
 }
 
 template <class T>
-T MFCC<T>::compH(int k, int m){
+T MFCC<T>::compH(size_t k, size_t m){
     T c=0;
     if (freq(k)<freq_c(m-1)){
         c=0;
