+++ BeatTheBox/Segment.h
@@ -19,29 +19,29 @@ class ISegmentOwner; // forward-declaration
 template <class T>
 class Segment{
 private: 
-    int _startDelta;
-    int _stopDelta;
+    size_t _startDelta;
+    size_t _stopDelta;
 public:
     Segment(ISegmentOwner<T>& owner, double sr);
     void init();
-    int getSegmentStartDelta();
-    int getSegmentStopDelta();
+    size_t getSegmentStartDelta();
+    size_t getSegmentStopDelta();
 
-    void findSegment(std::vector<T>* signal, int onset, std::vector<T> *result);
+    void findSegment(std::vector<T>* signal, size_t onset, std::vector<T> *result);
     
-    static int getStart(std::vector<T>* arr, int onset, int winSize);
-    static int getStop(std::vector<T>* arr, int onset, int winSize);
+    static size_t getStart(std::vector<T>* arr, size_t onset, int winSize);
+    static size_t getStop(std::vector<T>* arr, size_t onset, int winSize);
     
     void pushSample(T s, bool isOnset);
     
     
                         
-    int _signalLength;
+    size_t _signalLength;
     std::vector<T>* _signal;
 	// current position in the signal
-    int _signalPos;
+    size_t _signalPos;
 	// index of detected onset of the signal
-    int _onset;
+    size_t _onset;
     
     // flag signalling whether an onset has ever been detected
 	// will be set to and remain true, when the first onset is detected
